requirements.txt
# Core Dependencies
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
python-dotenv==0.19.0

# AI/ML Dependencies
transformers==4.11.3
torch==1.9.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2

# Database
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
alembic==1.7.1

# Message Queue
pika==1.2.0

# Testing
pytest==6.2.5
pytest-asyncio==0.15.1

# Monitoring
prometheus-client==0.11.0

# Utilities
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5 